requirements.txt
django
requests
django_extensions
django-markdown-deux
gunicorn
django-heroku
whitenoise
django-sass-processor
python-slugify
django-meta
django-htmlmin
django-debug-toolbar
django-reversion
django-reversion-compare